gonna change findItemById in ItemService

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/ProductService.java
@@ -49,13 +49,13 @@ public class ProductService implements IProductService{
 
     @Override
     public Product getProductById(long request) {
-        return this.repository.findAllById((long) request);
+        return this.repository.findAllById( request);
 
     }
 
     @Override
     public Boolean doesProductExist(long request) {
-        return this.repository.existsById((long) request);
+        return this.repository.existsById( request);
     }
 
     @Override
